+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/security/ClientConfiguration.java
@@ -1,0 +1,10 @@
+package org.openprovenance.prov.service.security;
+
+import org.pac4j.core.client.Client;
+
+public interface ClientConfiguration {
+    String DIRECT_BEARER_AUTH_CLIENT = "DirectBearerAuthClient";
+    String KEYCLOAK_OIDC_CLIENT = "KeycloakOidcClient";
+
+    Client configureClient();
+}